requirements.txt
requests>=2.31
pandas>=2.0
plotly>=5.18
python-dotenv>=1.0
pyarrow>=14.0
